refactor(frontend): move documents provider into documents presentation layer
