+++ frontend/src/modules/documents/index.ts
@@ -1,3 +1,4 @@
 export * from "./application";
 export * from "./domain";
 export * from "./infrastructure";
+export * from "./presentation";
