+++ frontend/src/modules/documents/presentation/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./document-provider";
